Improve PBG3 parser accuracy
