+++ config/implemented.csv
@@ -90,6 +90,7 @@ IPbg3Parser::ReadMagic
 IPbg3Parser::ReadString
 Pbg3Parser::Pbg3Parser
 Pbg3Parser::OpenArchive
+Pbg3Parser::Close
 Pbg3Parser::ReadBit
 Pbg3Parser::ReadInt
 Pbg3Parser::ReadByteAssumeAligned
@@ -97,6 +98,7 @@ Pbg3Parser::SeekToOffset
 Pbg3Parser::SeekToNextByte
 Pbg3Parser::ReadByteAlignedData
 Pbg3Parser::GetLastWriteTime
+Pbg3Parser::ReadByte
 Pbg3Parser::~Pbg3Parser
 Pbg3Archive::Pbg3Archive
 Pbg3Archive::ParseHeader
