+++ src/pbg3/Pbg3Parser.cpp
@@ -1,116 +1,37 @@
 #include "Pbg3Parser.hpp"
 
-void IPbg3Parser::Reset()
+Pbg3Parser::Pbg3Parser() : IPbg3Parser(), FileAbstraction()
 {
-    this->bitIdxInCurByte = 128;
-    this->offsetInFile = 0;
-    this->fileSize = 0;
-    this->curByte = 0;
-    this->crc = 0;
-}
-
-u32 IPbg3Parser::ReadVarInt()
-{
-    u32 res = 0;
-    i32 varintHdr = 0;
-
-    if (this->ReadBit())
-    {
-        varintHdr = 2;
-    }
-    if (this->ReadBit())
-    {
-        varintHdr |= 1;
-    }
-
-    u32 intLen;
-    switch (varintHdr)
-    {
-    case 0:
-        intLen = 0x80;
-        break;
-    case 1:
-        intLen = 0x8000;
-        break;
-    case 2:
-        intLen = 0x800000;
-        break;
-    case 3:
-        intLen = 0x80000000;
-        break;
-    default:
-        // TODO: There's probably a way to match without goto, but
-        // I can't figure it out... a simple `return 0;` won't share
-        // the function epilogue with the other return res.
-        goto end;
-    }
-
-    do
-    {
-        if (this->ReadBit())
-        {
-            res |= intLen;
-        }
-        intLen >>= 1;
-    } while (intLen != 0);
-end:
-    return res;
-}
-
-u32 IPbg3Parser::ReadMagic()
-{
-    u32 b0 = this->ReadInt(8);
-    u32 b1 = b0 + (this->ReadInt(8) << 8);
-    u32 b2 = b1 + (this->ReadInt(8) << 16);
-    u32 b3 = b2 + (this->ReadInt(8) << 24);
-
-    return b3;
-}
-
-u32 IPbg3Parser::ReadString(char *out, u32 maxSize)
-{
-    if (out == NULL)
-        return FALSE;
-
-    for (u32 idx = 0; idx < maxSize; idx++)
-    {
-        out[idx] = this->ReadInt(8);
-        if (out[idx] == '\0')
-        {
-            return TRUE;
-        }
-    }
-
-    return FALSE;
-}
-
-Pbg3Parser::Pbg3Parser() : IPbg3Parser(), fileAbstraction()
-{
-    this->Reset();
 }
 
 i32 Pbg3Parser::OpenArchive(char *path)
 {
-    this->fileAbstraction.Close();
+    this->Close();
     this->Reset();
-    if (this->fileAbstraction.Open(path, "r") == FALSE)
+    if (this->Open(path, "r") == FALSE)
     {
         return FALSE;
     }
-    this->fileSize = this->fileAbstraction.GetSize();
+    this->fileSize = this->GetSize();
     return TRUE;
+}
+
+void Pbg3Parser::Close()
+{
+    FileAbstraction::Close();
+    this->Reset();
 }
 
 i32 Pbg3Parser::ReadBit()
 {
-    if (!this->fileAbstraction.HasNonNullHandle())
+    if (!this->HasNonNullHandle())
     {
         return FALSE;
     }
 
     if (this->bitIdxInCurByte == 0x80)
     {
-        this->curByte = this->fileAbstraction.ReadByte();
+        this->curByte = this->ReadByte();
         if (this->curByte == -1)
         {
             return FALSE;
@@ -133,7 +54,7 @@ u32 Pbg3Parser::ReadInt(u32 numBitsAsPowersOf2)
     u32 remainingBits = 1 << (numBitsAsPowersOf2 - 1);
     u32 result = 0;
 
-    if (!this->fileAbstraction.HasNonNullHandle())
+    if (!this->HasNonNullHandle())
     {
         return 0;
     }
@@ -142,7 +63,7 @@ u32 Pbg3Parser::ReadInt(u32 numBitsAsPowersOf2)
     {
         if (this->bitIdxInCurByte == 0x80)
         {
-            this->curByte = this->fileAbstraction.ReadByte();
+            this->curByte = this->ReadByte();
             if (this->curByte == -1)
             {
                 return FALSE;
@@ -173,7 +94,7 @@ u8 Pbg3Parser::ReadByteAssumeAligned()
         this->offsetInFile += 1;
     }
 
-    return this->fileAbstraction.ReadByte();
+    return FileAbstraction::ReadByte();
 }
 
 i32 Pbg3Parser::SeekToOffset(u32 fileOffset)
@@ -188,7 +109,7 @@ i32 Pbg3Parser::SeekToOffset(u32 fileOffset)
         return FALSE;
     }
 
-    if (this->fileAbstraction.Seek(fileOffset, FILE_BEGIN) == FALSE)
+    if (this->Seek(fileOffset, FILE_BEGIN) == FALSE)
     {
         return FALSE;
     }
@@ -200,7 +121,7 @@ i32 Pbg3Parser::SeekToOffset(u32 fileOffset)
 
 i32 Pbg3Parser::SeekToNextByte()
 {
-    if (!this->fileAbstraction.HasNonNullHandle())
+    if (!this->HasNonNullHandle())
     {
         return FALSE;
     }
@@ -217,23 +138,28 @@ i32 Pbg3Parser::ReadByteAlignedData(u8 *data, u32 bytesToRead)
     u32 numBytesRead;
 
     this->SeekToNextByte();
-    return this->fileAbstraction.Read(data, bytesToRead, &numBytesRead);
+    return this->Read(data, bytesToRead, &numBytesRead);
 }
 
 i32 Pbg3Parser::GetLastWriteTime(LPFILETIME lastWriteTime)
 {
     // Yes, this is comparing against INVALID_HANDLE_VALUE instead of NULL. Why?
     // Unclear.
-    if (!this->fileAbstraction.HasValidHandle())
+    if (!this->HasValidHandle())
     {
         return FALSE;
     }
 
     // EWWWW abstraction violation much? (Maybe this is an inlined function?)
-    return this->fileAbstraction.GetLastWriteTime(lastWriteTime);
+    return this->GetLastWriteTime(lastWriteTime);
+}
+
+i32 Pbg3Parser::ReadByte()
+{
+    return this->ReadByteAssumeAligned();
 }
 
 Pbg3Parser::~Pbg3Parser()
 {
-    this->fileAbstraction.Close();
+    this->Close();
 }
