+++ src/pbg3/Pbg3Parser.hpp
@@ -1,32 +1,10 @@
 #pragma once
 
 #include "FileAbstraction.hpp"
+#include "IPbg3Parser.hpp"
 #include "inttypes.hpp"
 
-class IPbg3Parser
-{
-  public:
-    void Reset();
-    u32 ReadVarInt();
-    u32 ReadMagic();
-    u32 ReadString(char *out, u32 maxSize);
-    virtual i32 ReadBit() = 0;
-    virtual u32 ReadInt(u32 numBitsAsPowersOf2) = 0;
-    virtual u8 ReadByteAssumeAligned() = 0;
-    virtual i32 SeekToOffset(u32 fileOffset) = 0;
-    virtual i32 SeekToNextByte() = 0;
-    virtual i32 ReadByteAlignedData(u8 *data, u32 bytesToRead) = 0;
-    virtual i32 GetLastWriteTime(LPFILETIME lastWriteTime) = 0;
-
-  protected:
-    u32 offsetInFile;
-    u32 fileSize;
-    u32 curByte;
-    u8 bitIdxInCurByte;
-    u32 crc;
-};
-
-class Pbg3Parser : public IPbg3Parser
+class Pbg3Parser : public IPbg3Parser, public FileAbstraction
 {
   public:
     Pbg3Parser();
@@ -39,9 +17,9 @@ class Pbg3Parser : public IPbg3Parser
     i32 ReadByteAlignedData(u8 *data, u32 bytesToRead);
     i32 GetLastWriteTime(LPFILETIME lastWriteTime);
 
-    ~Pbg3Parser();
+    void Close();
+    i32 ReadByte();
 
-  private:
-    FileAbstraction fileAbstraction;
+    ~Pbg3Parser();
 };
 C_ASSERT(sizeof(Pbg3Parser) == 0x24);
